fend commit

+++ base/views.py
@@ -1,9 +1,14 @@
 from multiprocessing import context
+from django.contrib import messages
+from django.contrib.auth.models import User
+from django.contrib.auth import authenticate,login,logout
 from django.shortcuts import render , redirect
+from django.contrib.auth.decorators import login_required
 from django.http import HttpResponse, HttpResponseForbidden
-
+from django.db.models import Q
 from .forms import RoomForm
-from .models import Room
+from .models import Room, Topic
+
 # Create your views here.
 
 # rooms = [
@@ -25,17 +30,50 @@ from .models import Room
 
 # Functions for rendering
 
+def loginPage(request):
+        if request.user.is_authenticated:
+                return redirect('home')
+        if request.method == 'POST':
+                username=request.POST.get('username')
+                password=request.POST.get('password')
+
+                try:
+                        user=User.objects.get(username=username)
+                except:
+                        messages.error(request, 'User Not Fount.')
+
+                user = authenticate(request,username=username,password=password)
+
+                if user is not None:
+                        login(request,user)
+                        return redirect('home')
+                else:
+                        messages.error(request, 'Username or Password Not Exists.')
+
+        context={}
+        return(render(request, 'base/login_register.html', context))
+
+def logoutUser(request):
+        logout(request)
+        return redirect('home')
 
 def home(request):
-        rooms = Room.objects.all()
-        return render(request, 'home.html', {'rooms': rooms})
-
+        q=request.GET.get("q") if request.GET.get("q") != None else ''
+        rooms = Room.objects.filter(
+                Q(topic__name__icontains=q) |
+                Q(name__icontains=q) |
+                Q(description__icontains=q)
+                )
+        topics = Topic.objects.all()
+        room_count=rooms.count()
+        return render(request, 'home.html', {'rooms': rooms,'topics':topics, 'room_count':room_count })
 
 def room(request,pk):
     room = Room.objects.get(id=pk)
     context={'room': room}
     return render(request, 'room.html',context)
 
+@login_required(login_url='login')
 def create_room(request):
         form = RoomForm()
         if request.method == 'POST':
@@ -47,11 +85,12 @@ def create_room(request):
         context={'form': form}
         return(render(request, 'base/room_form.html', context))
 
-
-
+@login_required(login_url='login')
 def update_room(request,pk):
         room = Room.objects.get(id=pk)
         form = RoomForm(instance=room)
+        if request.user != room.host:
+                return HttpResponse("You are not allowed here")
         if request.method == 'POST':
                 form = RoomForm(request.POST, instance=room)
                 if form.is_valid():
@@ -59,9 +98,14 @@ def update_room(request,pk):
                         return redirect('home')
         context={'form': form}
         return render(request, 'base/room_form.html',context)
+@login_required(login_url='login')
 
 def delete_room(request,pk):
         room = Room.objects.get(id=pk)
+
+        if request.user != room.host:
+                return HttpResponse("You are not allowed here")
+
         if request.method=='POST':
                 room.delete()
                 return redirect('home')
